Move associative containers to subdirectory
