+++ libcxx/test/benchmarks/containers/associative/associative_container_benchmarks.h
@@ -16,7 +16,7 @@
 #include <vector>
 
 #include "benchmark/benchmark.h"
-#include "../GenerateInput.h"
+#include "../../GenerateInput.h"
 
 namespace support {
