+++ libcxx/test/benchmarks/containers/associative/multimap.bench.cpp
@@ -6,30 +6,29 @@
 //
 //===----------------------------------------------------------------------===//
 
-// REQUIRES: std-at-least-c++26
+// UNSUPPORTED: c++03, c++11, c++14, c++17
 
-#include <flat_map>
+#include <map>
 #include <string>
-#include <utility>
 
 #include "associative_container_benchmarks.h"
-#include "../GenerateInput.h"
+#include "../../GenerateInput.h"
 #include "benchmark/benchmark.h"
 
 template <class K, class V>
-struct support::adapt_operations<std::flat_map<K, V>> {
-  using ValueType = typename std::map<K, V>::value_type;
-  using KeyType   = typename std::map<K, V>::key_type;
+struct support::adapt_operations<std::multimap<K, V>> {
+  using ValueType = typename std::multimap<K, V>::value_type;
+  using KeyType   = typename std::multimap<K, V>::key_type;
   static ValueType value_from_key(KeyType const& k) { return {k, Generate<V>::arbitrary()}; }
   static KeyType key_from_value(ValueType const& value) { return value.first; }
 
-  using InsertionResult = std::pair<typename std::flat_map<K, V>::iterator, bool>;
-  static auto get_iterator(InsertionResult const& result) { return result.first; }
+  using InsertionResult = typename std::multimap<K, V>::iterator;
+  static auto get_iterator(InsertionResult const& result) { return result; }
 };
 
 int main(int argc, char** argv) {
-  support::associative_container_benchmarks<std::flat_map<int, int>>("std::flat_map<int, int>");
-  support::associative_container_benchmarks<std::flat_map<std::string, int>>("std::flat_map<std::string, int>");
+  support::associative_container_benchmarks<std::multimap<int, int>>("std::multimap<int, int>");
+  support::associative_container_benchmarks<std::multimap<std::string, int>>("std::multimap<std::string, int>");
 
   benchmark::Initialize(&argc, argv);
   benchmark::RunSpecifiedBenchmarks();
